Separate source. Optimize runner performance. Ignore if paired swift source file is not exist.
