+++ IBAnalyzer/Helpers/Collection+Has.swift
@@ -1,0 +1,15 @@
+//
+//  Collection+Has.swift
+//  IBAnalyzer
+//
+//  Created by USER on 14/10/2019.
+//  Copyright © 2019 Arkadiusz Holko. All rights reserved.
+//
+
+import Foundation
+
+extension Collection {
+    var hasElement: Bool {
+        return !isEmpty
+    }
+}
